Adding helper method and removing duplicated code
