+++ .gitattributes
@@ -1,2 +1,2 @@
 test/node/parsing/* linguist-generated
-test/node/parsing/IDLFragmentExtractor-test.js linguist-generated=false
+test/node/parsing/*.js linguist-generated=false
