+++ test/node/parsing/IDLFragmentExtractor-test.js
@@ -6,42 +6,13 @@
 describe('IDLFragmentExtractor', function() {
   var HTMLFileContents;
   var IDLFragmentExtractor;
-
-  function cmpTest(testName, testDirectory, numExpectedIDLFragments, raw) {
-    var fs = require('fs');
-    var dir = `${testDirectory}/${raw ? 'spec-raw' : 'spec'}.html`;
-    var spec = fs.readFileSync(dir).toString();
-    var htmlFile = HTMLFileContents.create({
-      url: dir, // Setting URL to dir for testing purposes only.
-      timestamp: new Date(),
-      contents: spec,
-    });
-
-    var extractor = IDLFragmentExtractor.create({file: htmlFile});
-    var idlFragments = extractor.idlFragments;
-
-    // Determine the number of fragments that were found.
-    expect(idlFragments.length).toBe(numExpectedIDLFragments);
-
-    fs.readdir(testDirectory, function(err, files) {
-      // Go through each of the expected results in the folder.
-      files.forEach(function(filename) {
-        var testNum = Number(filename);
-
-        if (!isNaN(testNum) && testNum < idlFragments.length) {
-          var path = `${testDirectory}/${filename}`;
-          var expectedContent = fs.readFileSync(path).toString();
-          expect(idlFragments[testNum].trim()).toBe(expectedContent.trim());
-        } else if (filename !== 'spec.html' && filename !== 'spec-raw.html') {
-          fail(`File ${filename} was not used in ${testName} spec test`);
-        }
-      });
-    });
-  }
+  var cmpTest;
 
   beforeEach(function() {
     HTMLFileContents = foam.lookup('org.chromium.webidl.HTMLFileContents');
     IDLFragmentExtractor = foam.lookup('org.chromium.webidl.IDLFragmentExtractor');
+    cmpTest = global.idlFragmentExtractorTest.bind(this,
+      HTMLFileContents, IDLFragmentExtractor);
   });
 
   it('should parse a pre tag with no content', function() {
